requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
psutil==5.9.0
PyNaCl==1.5.0
ffmpeg-python==0.2.0